chore (refs T22614): Move Plugins, VueConfig & VueEvents to client/js.
