+++ client/js/lib/core/ExtendedVueEvents.js
